core: Add (most) PathVec methods to PathRef
